refactor: extract discord_integration.py from __init__.py
Move Discord Rich Presence setup and hook wiring (~30 lines)
into discord_integration.py with a single setup_discord_hooks() call.
